Issue 154: PointedColorFloodFill causing Stack Overflow
Changing implementation of PointedColorFloodFill to use queue instead of stack, so the image processing routine will not generate stack overflow any more. It is speed is a bit reduced as a result of this, but it may get noticeable only when processing thousands of images (depends on image size of course).